backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
pywebpush
python-dotenv